% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex-methods.R
\name{as.regex}
\alias{as.regex}
\alias{is.regex}
\title{Convert or test for regex objects}
\usage{
as.regex(x)

is.regex(x)
}
\arguments{
\item{x}{An object to test or convert.}
}
\value{
\code{as.regex} returns the inputs object, with class
\code{c("regex", "character")}.
\code{is.regex} returns \code{TRUE} when the input inherits from class
\code{"regex"} and \code{FALSE} otherwise.
}
\description{
\code{as.regex} gives objects the class \code{"regex"}. \code{is.regex}
tests for objects of class \code{"regex"}.
}
\examples{
x <- as.regex("month.abb")
is.regex(x)
}
