% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backreferences.R
\docType{data}
\name{Backreferences}
\alias{Backreferences}
\alias{REF1}
\alias{REF2}
\alias{REF3}
\alias{REF4}
\alias{REF5}
\alias{REF6}
\alias{REF7}
\alias{REF8}
\alias{REF9}
\title{Backreferences}
\format{An object of class \code{regex} (inherits from \code{character}) of length 1.}
\usage{
REF1

REF2

REF3

REF4

REF5

REF6

REF7

REF8

REF9
}
\description{
Backreferences for replacement operations.
}
\examples{
REF1
REF2
# and so on, up to
REF9

sub("a(b)c(d)", REF1 \%R\% REF2, "abcd")
}
\references{
\url{http://www.regular-expressions.info/backref.html} and
\url{http://www.rexegg.com/regex-capture.html}
}
\keyword{datasets}

