% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture.R
\name{capture}
\alias{capture}
\alias{engroup}
\alias{group}
\alias{token}
\title{Capture a token, or not}
\usage{
capture(x)

group(x)

token(x)

engroup(x, capture)
}
\arguments{
\item{x}{A character vector.}

\item{capture}{Logical If \code{TRUE}, call \code{capture}; if \code{FALSE},
call \code{group}.}
}
\value{
A character vector representing part or all of a regular expression.
}
\description{
Create a token to capture or not.
}
\examples{
x <- "foo"
capture(x)
group(x)
}
\references{
\url{http://www.regular-expressions.info/brackets.html}
}

