% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/grouping-and-repetition.R
\name{char_class}
\alias{char_class}
\alias{negate_and_group}
\alias{negated_char_class}
\title{A range or char_class of characters.}
\usage{
char_class(...)

negated_char_class(...)

negate_and_group(...)
}
\arguments{
\item{...}{Character vectors.}
}
\value{
A character vector representing part or all of a regular expression.
}
\description{
Match a range or char_class of characters.
}
\examples{
char_class(LOWER, "._")
negated_char_class(LOWER, "._")

# Usage
(rx_odd <- char_class(1, 3, 5, 7, 9))
stringi::stri_detect_regex((1:10) ^ 2, rx_odd)
}
\references{
\url{http://www.regular-expressions.info/charclass.html}
}

