% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatenation.R
\name{Concatenation}
\alias{Concatenation}
\alias{\%R\%}
\alias{\%c\%}
\title{Combine strings together}
\usage{
x \%c\% y

x \%R\% y
}
\arguments{
\item{x}{A character vector.}

\item{y}{A character vector.}
}
\value{
A character vector representing part or all of a regular expression.
}
\description{
Operator equivalent of \code{regex}.
}
\note{
\code{\%c\%} was the original operator for this ('c' for
'concatenate').  This is hard work to type on a QWERTY keyboard
though, so it has been replaced with \code{\%R\%}.
}
\examples{
# Notice the recycling
letters \%R\% month.abb
}
\seealso{
\code{\link{regex}}, \code{\link[base]{paste}}
}

