% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{exactly}
\alias{exactly}
\title{Make a regex exact}
\usage{
exactly(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A character vector representing part or all of a regular expression.
}
\description{
Makes a regex exact: that is, it must contain the whole string, not just part
of it.
}
\examples{
exactly("#" \%R\% hex_digit(6))
}

