% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex-methods.R
\name{format.regex}
\alias{format.regex}
\alias{print.regex}
\title{Print or format regex objects}
\usage{
\method{format}{regex}(x, ...)

\method{print}{regex}(x, ...)
}
\arguments{
\item{x}{A regex object.}

\item{...}{Passed from other format methods.  Currently ignored.}
}
\value{
\code{format.regex} returns a character vector. \code{print.regex}
is invoked for the side effect of printing the regex object.
}
\description{
Prints/formats objects of class \code{regex}.
}
\examples{
group(1:5)
lookahead(1:5)
}

