% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{literal}
\alias{literal}
\title{Treat part of a regular expression literally.}
\usage{
literal(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A character vector representing part or all of a regular expression.
}
\description{
Treats its contents as literal characters.
}
\examples{
literal(digit(1, 3))
}

