% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookaround.R
\name{lookahead}
\alias{lookahead}
\alias{lookbehind}
\alias{negative_lookahead}
\alias{negative_lookbehind}
\title{Lookaround}
\usage{
lookahead(x)

negative_lookahead(x)

lookbehind(x)

negative_lookbehind(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A character vector representing part or all of a regular expression.
}
\description{
Zero length matching.
}
\examples{
x <- "foo"
lookahead(x)
negative_lookahead(x)
lookbehind(x)
negative_lookbehind(x)
}
\references{
\url{http://www.regular-expressions.info/lookaround.html} and
\url{http://www.rexegg.com/regex-lookarounds.html}
}

