% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex-methods.R
\name{regex}
\alias{regex}
\title{Create a regex}
\usage{
regex(...)
}
\arguments{
\item{...}{Passed to \code{paste0}.}
}
\value{
An object of class \code{regex}.
}
\description{
Creates a regex object.
}
\note{
This works like \code{paste0}, but the returns value has class
\code{c("regex", "character")}.
}
\seealso{
\code{\link[base]{paste0}}
as.regex(month.abb)
regex(letters[1:5], "?")
}

