% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouping-and-repetition.R
\name{repeat_in_class}
\alias{repeat_in_class}
\title{Engine for grouping and repeating classes}
\usage{
repeat_in_class(x, lo, hi, char_class)
}
\arguments{
\item{x}{A string.}

\item{lo}{A non-negative integer. Minimum number of repeats, when grouped.}

\item{hi}{positive integer. Maximum number of repeats, when grouped.}

\item{char_class}{\code{TRUE} or \code{FALSE}. Should the class be grouped?}
}
\value{
A character vector representing part or all of a regular expression.
}
\description{
An engine for the class functions that implicited calls \code{char_class} and
\code{repeated} for certain input combinations.
}

