% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{whole_word}
\alias{whole_word}
\title{Match a whole word}
\usage{
whole_word(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A character vector representing part or all of a regular expression.
}
\description{
Wraps a regex in word boundary tokens.
}
\examples{
# A hex color
(rx <- "cat")
(rx_whole_word <- whole_word(rx))

# Usage
stringi::stri_detect_regex(c("cat", "tomcat", "catastrophe"), rx)
stringi::stri_detect_regex(c("cat", "tomcat", "catastrophe"), rx_whole_word)
}
\seealso{
\code{\link{WordBoundaries}}, \code{\link{exactly}}
}

