% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookaround.R
\name{lookahead}
\alias{lookahead}
\alias{lookbehind}
\alias{negative_lookahead}
\alias{negative_lookbehind}
\title{Lookaround}
\usage{
lookahead(x)

negative_lookahead(x)

lookbehind(x)

negative_lookbehind(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A character vector representing part or all of a regular expression.
}
\description{
Zero length matching.
}
\note{
Lookbehind is not supported by R's PRCE engine.  Use R's Perl engine
or \code{stringi}/\code{stringr}'s ICU engine.
}
\examples{
x <- "foo"
lookahead(x)
negative_lookahead(x)
lookbehind(x)
negative_lookbehind(x)

# Usage
x <- c("mozambique", "qatar", "iraq")
(rx_neg_class <- "q" \%R\% negated_char_class("u"))
(rx_neg_lookahead <- "q" \%R\% negative_lookahead("u"))
stringi::stri_detect_regex(x, rx_neg_class)
stringi::stri_detect_regex(x, rx_neg_lookahead)
}
\references{
\url{http://www.regular-expressions.info/lookaround.html} and
\url{http://www.rexegg.com/regex-lookarounds.html}
}

