% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{WordBoundaries}
\alias{BOUNDARY}
\alias{NOT_BOUNDARY}
\alias{WordBoundaries}
\title{Word boundaries}
\format{An object of class \code{regex} (inherits from \code{character}) of length 1.}
\usage{
BOUNDARY

NOT_BOUNDARY
}
\description{
Match a word boundary.
}
\examples{
BOUNDARY
NOT_BOUNDARY

# Usage
x <- c("the catfish", "the tomcat")
rx <- BOUNDARY \%R\% "cat"
stringi::stri_detect_regex(x, rx)
}
\references{
\url{http://www.regular-expressions.info/wordboundaries.html} and
\url{http://www.rexegg.com/regex-boundaries.html}
}
\seealso{
\code{\link{ALPHA}}, \code{\link{BACKSLASH}}, \code{\link{START}}
}
\keyword{datasets}

